feat(energy): share continuous nominal model with route optimization

+++ src/main/java/com/navio/mobilityandevservice/domain/optimization/EvPlanOperation.java
@@ -8,9 +8,9 @@ public record EvPlanOperation(
         String beforeItemId,
         int sequence,
         EvChargerResponse charger,
-        int estimatedChargeMinutes,
-        int arrivalSocPct,
-        int departureSocPct,
+        double estimatedChargeMinutes,
+        double arrivalSocPct,
+        double departureSocPct,
         double detourKm,
         String reason
 ) {
